[5367] * [new sql update: 5367_mangos_spell_affect.sql] Implement rogue talent that increase some spells damage and prevent dodge for finishing moves. * Now Unit::MeleeMissChanceCalc and Unit::SpellMissChanceCalc return float instead int32(percent*100) like other similar functions.

+++ src/game/Pet.cpp
@@ -598,7 +598,7 @@ void Pet::ModifyLoyalty(int32 addvalue)
             --loyaltylevel;
             SetLoyaltyLevel(LoyaltyLevel(loyaltylevel));
             m_loyaltyPoints = GetStartLoyaltyPoints(loyaltylevel);
-            SetTP(m_TrainingPoints - getLevel());
+            SetTP(m_TrainingPoints - int32(getLevel()));
         }
         else
         {
